done all history fetching and display
