+++ apps/client/components/history/history-card.tsx
@@ -1,4 +1,5 @@
-import { Badge, Card, Group, Text } from '@mantine/core';
+'use client';
+import { Badge, Card, Group, Text, Image } from '@mantine/core';
 
 interface HistoryCardProps {
   status: string;
@@ -6,6 +7,7 @@ interface HistoryCardProps {
   title: string;
   description: string;
   awarded: string;
+  imageUrl: string;
 }
 
 export default function HistoryCard({
@@ -14,7 +16,10 @@ export default function HistoryCard({
   title,
   description,
   awarded,
+  imageUrl,
 }: HistoryCardProps) {
+  const statusColor = status.toLowerCase() === 'won' ? 'green' : 'red';
+
   return (
     <Card
       padding="lg"
@@ -23,8 +28,18 @@ export default function HistoryCard({
       className="shadow-sm hover:shadow-lg transition duration-300 ease-in-out"
       pos="relative"
     >
-      <Group justify="space-between">
-        <Badge color="green">{status}</Badge>
+      <Card.Section>
+        <Image
+          alt="Image"
+          height={160}
+          src={
+            imageUrl ||
+            'https://placehold.jp/24/3d4070/ffffff/1280x720.png?text=No%20Image'
+          }
+        />
+      </Card.Section>
+      <Group mt='xs' justify="space-between ">
+        <Badge color={statusColor}>{status}</Badge>
         <Badge color="blue" variant="light" p="sm">
           Date Posted : {datePosted}
         </Badge>
@@ -32,10 +47,10 @@ export default function HistoryCard({
       <Text size="xl" mt="sm" fw={600}>
         {title}
       </Text>
-      <p className="text-md h-15 overflow-y-auto">{description}</p>
+      <p className="text-md max-h-16  overflow-y-auto">{description}</p>
       <Group justify="space-between">
         <Text fw={500}>
-          AWARDED : <span className='font-extrabold'>{awarded}</span>
+          AWARDED : <span className="font-extrabold">{awarded}</span>
         </Text>
         <Badge color="gray" variant="light" radius="sm">
           HACKATHON
