+++ apps/client/components/history/history-page.tsx
@@ -9,16 +9,19 @@ export default async function HistoryPage() {
       <div className="grid grid-cols-3 gap-4">
         {data.map((prize) => {
           if (
-            prize.Awarded &&
-            prize.PrizeName &&
-            prize.DatePosted &&
-            prize.AwardedUSDe &&
-            prize.WinnersAmount
+            (prize.Awarded &&
+              // prize.WonRefunded &&
+              prize.PrizeName) ||
+            prize.DatePosted ||
+            prize.AwardedUSDe ||
+            prize.WinnersAmount // Changed this line
           ) {
+            const status = prize.WinnersAmount ? 'Won' : 'Not Won'; // Adjusted status based on WinnersAmount existence
             return (
               <HistoryCard
+                imageUrl={prize.Image}
                 key={prize.PrizeName}
-                status="Won"
+                status={status} // Passed the adjusted status here
                 datePosted={prize.DatePosted}
                 title={prize.PrizeName}
                 description={prize.WinnersAmount}
